fix: Add daily score calculation job for finding Post query
