+++ packages/velog-cron/src/common/plugins/globals/registerJobPlugin.ts
@@ -1,0 +1,35 @@
+import { PostJob } from '@jobs/PostJob'
+import { FastifyPluginCallback } from 'fastify'
+import { container } from 'tsyringe'
+
+const createJobPlugin: FastifyPluginCallback = (fastfiy, opts, done) => {
+  const postJob = container.resolve(PostJob)
+
+  const scoreCalculateJobInFiveMin = fastfiy.cron.createJob({
+    name: 'posts score calculation in every 5 minutes',
+    cronTime: '*/5 * * * *', // every 5 minutes
+    onTick: async () => {
+      if (postJob.isJobProgressing) return
+      postJob.start()
+      await postJob.calculateScore(0.5)
+      postJob.stop()
+    },
+  })
+
+  const scoreCalculateJobInDay = fastfiy.cron.createJob({
+    name: 'posts score calculation in every day',
+    cronTime: '0 6 * * *', // every day at 06:00 (6:00 AM)
+    onTick: async () => {
+      if (postJob.isJobProgressing) return
+      postJob.start()
+      await postJob.calculateScore(0.1)
+      postJob.stop()
+    },
+  })
+
+  scoreCalculateJobInFiveMin.start()
+  scoreCalculateJobInDay.start()
+  done()
+}
+
+export default createJobPlugin
