+++ packages/velog-cron/src/jobs/JobProgress.ts
@@ -1,0 +1,12 @@
+export class JobProgress {
+  private jobInProgress = false
+  public get isJobProgressing() {
+    return this.jobInProgress
+  }
+  public start() {
+    this.jobInProgress = true
+  }
+  public stop() {
+    this.jobInProgress = false
+  }
+}
